// src/auth/enums/permission.enum.ts
export enum AppPermission {
  MANAGE_UNITIES = 'manage_unities',
  MANAGE_MULTIMEDIA = 'manage_multimedia',
  MANAGE_LESSONS = 'manage_lessons',
  MANAGE_EXERCISES = 'manage_exercises',
  MANAGE_CULTURAL_ACHIEVEMENTS = 'manage_cultural_achievements',
  MANAGE_MISSION_TEMPLATES = 'manage_mission_templates',
  MANAGE_CONTENT = 'manage_content',
  MANAGE_REWARDS = 'manage_rewards',
  MANAGE_ACTIVITIES = 'manage_activities',
  // Add more granular permissions as needed
  CREATE_UNITY = 'create_unity',
  READ_UNITIES = 'read_unities',
  READ_UNITY = 'read_unity',
  UPDATE_UNITY = 'update_unity',
  DELETE_UNITY = 'delete_unity',
  TOGGLE_LOCK_UNITY = 'toggle_lock_unity',
  UPDATE_UNITY_POINTS = 'update_unity_points',
  UPLOAD_MULTIMEDIA = 'upload_multimedia',
  READ_MULTIMEDIA_LIST = 'read_multimedia_list',
  READ_MULTIMEDIA_DETAIL = 'read_multimedia_detail',
  DELETE_MULTIMEDIA = 'delete_multimedia',
  DOWNLOAD_MULTIMEDIA = 'download_multimedia',
  CREATE_CONTENT = 'create_content',
  READ_CONTENT_LIST = 'read_content_list',
  READ_CONTENT_DETAIL = 'read_content_detail',
  UPDATE_CONTENT = 'update_content',
  DELETE_CONTENT = 'delete_content',
  READ_CONTENT_BY_UNITY_AND_TOPIC = 'read_content_by_unity_and_topic',
  CREATE_LESSON = 'create_lesson',
  READ_LESSONS_LIST = 'read_lessons_list',
  READ_LESSON_DETAIL = 'read_lesson_detail',
  READ_LESSONS_BY_UNITY = 'read_lessons_by_unity',
  UPDATE_LESSON = 'update_lesson',
  DELETE_LESSON = 'delete_lesson',
  TOGGLE_LOCK_LESSON = 'toggle_lock_lesson',
  UPDATE_LESSON_POINTS = 'update_lesson_points',
  MARK_LESSON_COMPLETED = 'mark_lesson_completed',
  CREATE_EXERCISE = 'create_exercise',
  READ_EXERCISES_LIST = 'read_exercises_list',
  READ_EXERCISE_DETAIL = 'read_exercise_detail',
  UPDATE_EXERCISE = 'update_exercise',
  DELETE_EXERCISE = 'delete_exercise',
  CREATE_CULTURAL_ACHIEVEMENT = 'create_cultural_achievement',
  READ_CULTURAL_ACHIEVEMENTS_LIST = 'read_cultural_achievements_list',
  READ_CULTURAL_ACHIEVEMENT_DETAIL = 'read_cultural_achievement_detail',
  INITIALIZE_CULTURAL_ACHIEVEMENT_PROGRESS = 'initialize_cultural_achievement_progress',
  UPDATE_CULTURAL_ACHIEVEMENT_PROGRESS = 'update_cultural_achievement_progress',
  READ_USER_CULTURAL_ACHIEVEMENTS = 'read_user_cultural_achievements',
  READ_CULTURAL_ACHIEVEMENT_PROGRESS = 'read_cultural_achievement_progress',
  CREATE_MISSION_TEMPLATE = 'create_mission_template',
  READ_MISSION_TEMPLATES_LIST = 'read_mission_templates_list',
  READ_MISSION_TEMPLATE_DETAIL = 'read_mission_template_detail',
  UPDATE_MISSION_TEMPLATE = 'update_mission_template',
  CREATE_REWARD = 'create_reward',
  READ_REWARDS_LIST = 'read_rewards_list',
  AWARD_REWARD_TO_USER = 'award_reward_to_user',
  READ_USER_REWARDS = 'read_user_rewards',
  CONSUME_REWARD = 'consume_reward',
  CHECK_REWARD_STATUS = 'check_reward_status',
  CREATE_ACTIVITY = 'create_activity',
  READ_ACTIVITIES_LIST = 'read_activities_list',
  READ_ACTIVITY_DETAIL = 'read_activity_detail',
  READ_ACTIVITIES_BY_TYPE = 'read_activities_by_type',
  READ_ACTIVITIES_BY_DIFFICULTY = 'read_activities_by_difficulty',
  UPDATE_ACTIVITY = 'update_activity',
  DELETE_ACTIVITY = 'delete_activity',
  UPDATE_ACTIVITY_POINTS = 'update_activity_points',
}